copilot/content-generator/requirements.txt
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-monitor-opentelemetry==1.6.2
azure-search==1.0.0b2
azure-search-documents==11.5.2
dependency-injector==4.43.0
fastapi==0.115.3
llama-index==0.11.19
llama-index-core==0.11.19
llama-index-embeddings-azure-openai==0.2.5
llama-index-utils-workflow==0.2.2
opentelemetry-instrumentation-logging==0.48b0
pydantic==2.9.2
pyodbc==5.2.0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.1.1
uvicorn==0.32.0
holidays==0.68
PyPDF2==3.0.1